+++ changelog.md
@@ -1006,4 +1006,5 @@ Security in case of vulnerabilities.
 - dynamic menu container
 - clean HTML on deploy template (option -> html.clean=true (config.properties))
 - black and white image filter
+- macro to transform all links in : page:##name##